refactor: 整理代码风格与工具链配置，新增文件系统与终端功能
- 重构帧分配器、分页器、调度器等核心模块代码风格
- 新增控制台背景、启动LOGO、阻塞读行等终端功能
- 实现MorionFS文件系统与挂载层
- 新增能力句柄系统与相关系统调用
- 完善文档与Makefile配置

+++ kernel/src/arch/idt.rs
@@ -114,7 +114,7 @@ extern "x86-interrupt" fn page_fault_handler(
         crate::video::print_hex(stack_frame.instruction_pointer.as_u64());
         crate::video::println("");
         crate::video::print("error code: 0x");
-        crate::video::print_hex(error_code.bits() as u64);
+        crate::video::print_hex(error_code.bits());
         crate::video::println("");
         crate::video::print("rsp:        0x");
         crate::video::print_hex(stack_frame.stack_pointer.as_u64());
